move package from BEAR.Sunday

+++ apps/Sandbox/scripts/instance.php
@@ -10,15 +10,10 @@
  */
 namespace Sandbox;
 
-use Ray\Di\Injector;
-use Ray\Di\Container;
-use Ray\Di\Forge;
-use Ray\Di\ApcConfig;
-use Ray\Di\Annotation;
-use Ray\Di\Definition;
+use Ray\Di;
 use Doctrine\Common\Annotations\AnnotationReader;
 use BEAR\Sunday\Output\Console;
-use BEAR\Sunday\Web\SymfonyResponse;
+use BEAR\Package\Web\SymfonyResponse;
 use BEAR\Sunday\Exception\ExceptionHandler;
 
 // init
@@ -39,12 +34,21 @@ if ($app) {
 // create app object with  "run-mode" module
 $moduleName = __NAMESPACE__ . '\Module\\' . $mode . 'Module';
 
+// valid mode ?
+if (! class_exists($moduleName)) {
+    throw new \LogicException("Invalid configuration mode[{$mode}]");
+}
+
+// create injector with mode module
 $module = new $moduleName;
-$injector = new Injector(new Container(new Forge(new ApcConfig(new Annotation(new Definition, new AnnotationReader)))), new $module);
+$injector = new Di\Injector(new Di\Container(new Di\Forge(new Di\ApcConfig(new Di\Annotation(new Di\Definition, new AnnotationReader)))), new $module);
+
 // log binding info
 $log = (string)$injector;
 file_put_contents(dirname(__DIR__) . "/data/log/module.{$cacheKey}.log", $log);
 file_put_contents(dirname(__DIR__) . "/data/log/di.log.txt", $log);
+
+// crete application object
 try {
     $app = $injector->getInstance('BEAR\Sunday\Application\Context');
 } catch (\Exception $e) {
